+++ sys/GuiMenu.cpp
@@ -273,7 +273,7 @@ void GuiMenu_empty (GuiMenu me) {
 	@end
 #endif
 
-GuiMenu GuiMenu_createInWindow (GuiWindow window, const char32 *title, long flags) {
+GuiMenu GuiMenu_createInWindow (GuiWindow window, const char32 *title, uint32 flags) {
 	GuiMenu me = Thing_new (GuiMenu);
 	my d_shell = window;
 	my d_parent = window;
@@ -398,7 +398,7 @@ GuiMenu GuiMenu_createInWindow (GuiWindow window, const char32 *title, long flag
 	return me;
 }
 
-GuiMenu GuiMenu_createInMenu (GuiMenu supermenu, const char32 *title, long flags) {
+GuiMenu GuiMenu_createInMenu (GuiMenu supermenu, const char32 *title, uint32 flags) {
 	GuiMenu me = Thing_new (GuiMenu);
 	my d_shell = supermenu -> d_shell;
 	my d_parent = supermenu;
@@ -488,7 +488,7 @@ static gint button_press (GtkWidget *widget, GdkEvent *event)
 }
 #endif
 
-GuiMenu GuiMenu_createInForm (GuiForm form, int left, int right, int top, int bottom, const char32 *title, long flags) {
+GuiMenu GuiMenu_createInForm (GuiForm form, int left, int right, int top, int bottom, const char32 *title, uint32 flags) {
 	GuiMenu me = Thing_new (GuiMenu);
 	my d_shell = form -> d_shell;
 	my d_parent = form;
